data-enrichment/1-artefact-embedding/pyproject.toml
[tool.poetry]
name = "xcurator-artefact-embedding"
version = "0.2.2" # placeholder | version is defined in git tags | command: poetry version
description = "extracts clip text and visual embeddings."
authors = ["Sören Räuchle <sraeuchle@3pc.de>"]
license = "MIT"
readme = "README.md"

# https://pypi.org/project/poetry-dynamic-versioning/
[tool.poetry-dynamic-versioning]
enable = false
vcs = 'git'
style = "semver"
pattern = '(?x)^v*((?P<epoch>\d+)!)?(?P<base>\d+(\.\d+)*)' # 0.0.0 || v0.0.0 (git tag)

[tool.poetry.build]
generate-setup-file = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = { version = "2.1.0", source="pytorch-cu118" }
torchvision = { version = "0.16.0", source="pytorch-cu118" }
open-clip-torch="2.23.0"
multilingual-clip="1.0.10"
3pc-py-config = {path="./lib/3pc-py-config-1.1.1.tar.gz"}
overrides="7.4.0"
pydantic="2.5.1"
pandas="2.1.3"


[[tool.poetry.source]]
name = 'pytorch-cu118'
url = 'https://download.pytorch.org/whl/cu118'
priority = "supplemental"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.6"

[tool.poetry.group.test.dependencies]
pytest = "7.3.1"
pytest-asyncio = "0.21.0"
matplotlib = "3.7.1"

#######################################
#   COMMON Settings
#######################################

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend" # wrapper around poetry.core.masonry.api

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"